Bring core module extensions to core module
Then simplify dependencies
